requirements.txt
numpy
pandas

# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

# prevent bq 3+ db-dtypes package error when running training from bq

python-dotenv
google-cloud-storage

# API
fastapi
uvicorn

python-multipart
opencv-python
imageio
matplotlib
IPython
pytest-shutil
skia-python
dataclasses
Pillow
requires.io
